Initial fix to issue with initializing table metadata manager
