+++ crates/modelardb_manager/Cargo.toml
@@ -26,7 +26,6 @@ path = "src/main.rs"
 [dependencies]
 arrow.workspace = true
 arrow-flight.workspace = true
-dashmap.workspace = true
 futures.workspace = true
 modelardb_common = { path = "../modelardb_common" }
 object_store = { workspace = true, features = ["aws", "azure"] }
